refactor: split scheduling algorithms into separate modules (fcfs, sjf, stcf, rr, mlfq)
- extracted each algorithm into its own .c/.h file
- removed monolithic scheduler.c structure
- updated Makefile to include modular sources
- preserved existing functionality and outputs
- improved code maintainability and modularity
