+++ include/fcfs.h
@@ -1,0 +1,8 @@
+#ifndef FCFS_H
+#define FCFS_H
+
+#include "process.h"
+
+int schedule_fcfs(Process *processes, int n);
+
+#endif
