+++ include/mlfq.h
@@ -1,0 +1,8 @@
+#ifndef MLFQ_H
+#define MLFQ_H
+
+#include "process.h"
+
+int schedule_mlfq(Process *processes, int n);
+
+#endif
