+++ include/rr.h
@@ -1,0 +1,8 @@
+#ifndef RR_H
+#define RR_H
+
+#include "process.h"
+
+int schedule_rr(Process *processes, int n, int quantum);
+
+#endif
