+++ include/sjf.h
@@ -1,0 +1,8 @@
+#ifndef SJF_H
+#define SJF_H
+
+#include "process.h"
+
+int schedule_sjf(Process *processes, int n);
+
+#endif
