+++ include/stcf.h
@@ -1,0 +1,8 @@
+#ifndef STCF_H
+#define STCF_H
+
+#include "process.h"
+
+int schedule_stcf(Process *processes, int n);
+
+#endif
